# tms/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ScyllaX-TMS API in Go using Fiber framework",
        "title": "ScyllaX-TMS API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/mobile/pickup": {
            "post": {
                "description": "save pickup all to db.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pickup"
                ],
                "summary": "pickup all",
                "parameters": [
                    {
                        "description": "pickup all",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.PickUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.JsonSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonBadRequest"
                        }
                    },
                    "404": {
                        "description": "Data not found",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonInternalServerError"
                        }
                    }
                }
            }
        },
        "/mobile/pickup/partial": {
            "post": {
                "description": "save pickup partial to db.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pickup"
                ],
                "summary": "pickup partial",
                "parameters": [
                    {
                        "description": "pickup partial",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.PickupPartialRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.JsonSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonBadRequest"
                        }
                    },
                    "404": {
                        "description": "Data not found",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonInternalServerError"
                        }
                    }
                }
            }
        },
        "/mobile/pickup/skip": {
            "post": {
                "description": "save skip pickup to db.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skip pickup"
                ],
                "summary": "skip pickup",
                "parameters": [
                    {
                        "description": "skip pickup",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.SkipPickUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.JsonSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonBadRequest"
                        }
                    },
                    "404": {
                        "description": "Data not found",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/entity.JsonInternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.JsonBadRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "email": "email is required",
                        "username": "username is required"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "BAD REQUEST"
                },
                "trace_id": {
                    "type": "string",
                    "example": "dedc5250-5c20-48c9-9383-fac3ccff2679"
                }
            }
        },
...
        "entity.JsonInternalServerError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "errors": {
                    "type": "string",
                    "example": "error database or third party"
                },
                "status": {
                    "type": "string",
                    "example": "INTERNAL SERVER ERROR"
                },
                "trace_id": {
                    "type": "string",
                    "example": "dedc5250-5c20-48c9-9383-fac3ccff2679"
                }
            }
        },
        "entity.JsonNotFound": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "errors": {
                    "type": "string",
                    "example": "record not found"
                },
                "status": {
                    "type": "string",
                    "example": "NOT FOUND"
                },
                "trace_id": {
                    "type": "string",
                    "example": "dedc5250-5c20-48c9-9383-fac3ccff2679"
                }
            }
        },
        "entity.JsonSuccess": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "trace_id": {
                    "type": "string",
                    "example": "dedc5250-5c20-48c9-9383-fac3ccff2679"
                }
            }
        },
...
        "entity.PickUpRequest": {
            "type": "object",
            "required": [
                "current_time",
                "id"
            ],
            "properties": {
                "current_time": {
                    "type": "integer",
                    "example": 1727233938570
                },
                "id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1218
                    ]
                }
            }
        },
...
        "entity.PickupPartialBody": {
            "type": "object",
            "required": [
                "current_time",
                "outlet_id",
                "products",
                "shipment_no"
            ],
            "properties": {
                "current_time": {
                    "type": "integer",
                    "example": 1727233938570
                },
                "outlet_id": {
                    "type": "integer",
                    "example": 88
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Product"
                    }
                },
                "shipment_no": {
                    "type": "string",
                    "example": "DO202409250178"
                }
            }
        },
        "entity.PickupPartialRequest": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/entity.PickupPartialBody"
                }
            }
        },
        "entity.Product": {
            "type": "object",
            "required": [
                "id",
                "qty"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1224
                },
                "qty": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Quantity"
                    }
                }
            }
        },
...
        "entity.Quantity": {
            "type": "object",
            "required": [
                "unit_id"
            ],
            "properties": {
                "stock": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 2
                },
                "unit_id": {
                    "type": "string",
                    "example": "CTN"
                }
            }
        },
...
        "entity.SkipPickUpBody": {
            "type": "object",
            "required": [
                "id",
                "reason_id",
                "reason_name"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1218
                },
                "reason_id": {
                    "type": "integer",
                    "example": 1
                },
                "reason_name": {
                    "type": "string",
                    "example": "Reject"
                }
            }
        },
        "entity.SkipPickUpRequest": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.SkipPickUpBody"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}